Move editors in their diagram type package
